+++ internal/storage/sqlindexer.go
@@ -17,7 +17,6 @@ import (
 	"time"
 
 	"cloud.google.com/go/storage"
-	"github.com/hashicorp/golang-lru/v2/expirable"
 	"github.com/prometheus/client_golang/prometheus"
 	"go.elastic.co/apm/v2"
 	"go.uber.org/zap"
@@ -57,7 +56,7 @@ type SQLIndexer struct {
 
 	readPackagesBatchSize int
 
-	cache *expirable.LRU[string, []byte] // Cache for search results
+	afterUpdateHook func(ctx context.Context)
 }
 
 type IndexerOptions struct {
@@ -67,8 +66,8 @@ type IndexerOptions struct {
 	WatchInterval                time.Duration
 	Database                     database.Repository
 	SwapDatabase                 database.Repository
-	Cache                        *expirable.LRU[string, []byte] // Cache for search results
 	ReadPackagesBatchsize        int
+	AfterUpdateIndexHook         func(ctx context.Context)
 }
 
 func NewIndexer(logger *zap.Logger, storageClient *storage.Client, options IndexerOptions) *SQLIndexer {
@@ -84,8 +83,8 @@ func NewIndexer(logger *zap.Logger, storageClient *storage.Client, options Index
 		swapDatabase:          options.SwapDatabase,
 		label:                 fmt.Sprintf("storage-%s", options.PackageStorageEndpoint),
 		readPackagesBatchSize: defaultReadPackagesBatchSize,
-		cache:                 options.Cache,
 		cursor:                "init",
+		afterUpdateHook:       options.AfterUpdateIndexHook,
 	}
 
 	indexer.current = &indexer.database
@@ -300,10 +299,8 @@ func (i *SQLIndexer) swapDatabases(ctx context.Context, currentCursor string, nu
 	i.current, i.backup = i.backup, i.current
 	i.logger.Debug("Current database changed", zap.String("current.database.path", (*i.current).File(ctx)), zap.String("previous.database.path", (*i.backup).File(ctx)))
 
-	if i.cache != nil {
-		// Clear the cache after updating the index
-		// there could be new, updated or removed packages
-		i.cache.Purge()
+	if i.afterUpdateHook != nil {
+		i.afterUpdateHook(ctx)
 	}
 
 	metrics.StorageIndexerUpdateIndexSuccessTotal.Inc()